Finalize cache implementation, remove provider
