+++ src/main/kotlin/com/edelweiss/playerex/PlayerEXComponents.kt
@@ -14,12 +14,10 @@ import net.minecraft.entity.LivingEntity
 
 class PlayerEXComponents : EntityComponentInitializer, LevelComponentInitializer, WorldComponentInitializer {
     companion object {
-        val livingEntities: ComponentKey<LivingEntityComponent> = ComponentRegistry.getOrCreate(PlayerEXAPI.id("living_entities"), LivingEntityComponent::class.java)
         val playerEntities: ComponentKey<PlayerEntityComponent> = ComponentRegistry.getOrCreate(PlayerEXAPI.id("player_entities"), PlayerEntityComponent::class.java)
     }
 
     override fun registerEntityComponentFactories(registry: EntityComponentFactoryRegistry) {
-        registry.registerFor(LivingEntity::class.java, livingEntities) { entity -> LivingEntityComponent(entity) }
         registry.registerForPlayers(playerEntities) { player -> PlayerEntityComponent(player) }
     }
