+++ src/main/kotlin/com/edelweiss/playerex/cache/PlayerCacheKeys.kt
@@ -1,7 +1,7 @@
 package com.edelweiss.playerex.cache
 
-enum class PlayerCacheKeys(val key: String) {
-    UUID("uuid"),
-    Name("name"),
-    Keys("keys"),
+object PlayerCacheKeys {
+    const val UUID = "uuid"
+    const val NAME = "name"
+    const val KEYS = "keys"
 }
